requirements.txt
mysqlbinlog2blinker
gcloud
PyYAML